Make the tests work again.
Switching to Sqlite caused the tests to fail, because the connection string was
not included.

The tests now also use the intended 'GetDiscoveryDocumentAsync' function to
query the discovery endpoint. This also solves the issue with the 'IsError' not
catching all errors, in this case the 'GetDiscoveryDocumentAsync' fails and sets
'IsError' to true.